ensure id is found for zfs

+++ lib/getch/filesystem/zfs/encrypt/format.rb
@@ -5,7 +5,15 @@ module Getch
         class Format < Getch::FileSystem::Zfs::Encrypt::Device
           def initialize
             super
+            @log = Getch::Log.new()
             @state = Getch::States.new()
+            if ! @id
+              @log.info "Research pool id for #{@dev_root}..."
+              sleep 2 until Helpers::pool_id(@dev_root)
+              @id = Helpers::pool_id(@dev_root)
+              @boot_pool_name = "bpool-#{@id}"
+              @pool_name = "rpool-#{@id}"
+            end
             format
           end
 
@@ -13,6 +21,7 @@ module Getch
 
           def format
             return if STATES[:format]
+            raise "Error, no id found for #{@dev_root}." if ! @id
             system("mkfs.fat -F32 #{@dev_esp}") if @dev_esp
             system("mkswap -f #{@dev_swap}")
             zfs
